fix(be): fix logic of file buildingblock, modelbinder, insert blog

+++ src/src/Modules/Application/Blog.Domain.Application/Entities/Banner.cs
@@ -8,7 +8,7 @@ public class Banner : BaseEntityWithAudit
 {
     public required string PublicId { get; set; }
     public required string Url { get; set; }
-    public float Width { get; set; }
-    public float Height { get; set; }
+    public double Width { get; set; }
+    public double Height { get; set; }
     public string ETag { get; set; } = default!;
 }
